100 move all textcomponent styles to own files
