+++ src/Components/TextField/TextFieldFilled/TextFilledFilled.styles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+const styles = StyleSheet.create({
+  containerStyle: {
+    position: 'relative',
+  },
+  textField: {
+    textAlignVertical: 'bottom',
+    paddingTop: 16,
+    paddingHorizontal: 12,
+  },
+
+  filledInput: {
+    borderTopLeftRadius: 4,
+    borderTopRightRadius: 4,
+    backgroundColor: '#d9d9d9',
+  },
+});
+
+export default styles;
