+++ src/Components/TextField/TextFieldHelperText/TextFieldHelperText.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Text, StyleSheet } from 'react-native';
+import { Text } from 'react-native';
 import withTheme from '../../../Theme/withTheme';
+import styles from './TextFieldHelperText.styles';
 
 class TextFieldHelperText extends Component {
   static propTypes = {
@@ -27,17 +28,5 @@ class TextFieldHelperText extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  helper: {
-    color: 'rgba(0,0,0,0.6)',
-    height: 16,
-    marginLeft: 12,
-    zIndex: 10,
-    width: '100%',
-    position: 'absolute',
-    bottom: -20,
-  },
-});
 
 export default withTheme(TextFieldHelperText);
