+++ src/Components/TextField/TextFieldHelperText/TextFieldHelperText.styles.js
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  helper: {
+    color: 'rgba(0,0,0,0.6)',
+    height: 16,
+    marginLeft: 12,
+    zIndex: 10,
+    width: '100%',
+    position: 'absolute',
+    bottom: -20,
+  },
+});
+export default styles;
