+++ src/Components/TextField/TextFieldLabel/TextFieldLabel.styles.js
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  label: {
+    position: 'absolute',
+    zIndex: 10,
+  },
+});
+export default styles;
