+++ src/Components/TextField/TextFieldOutline/TextFieldOutline.styles.js
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  containerStyle: {},
+  textField: {
+    height: 56,
+    paddingHorizontal: 12,
+  },
+  outlinedInput: {
+    borderWidth: 1,
+    borderRadius: 4,
+  },
+});
+export default styles;
