+++ src/Components/TextField/TextFieldUnderline/TextFieldUnderline.styles.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+const styles = StyleSheet.create({
+  underline: {
+    position: 'absolute',
+    left: 0,
+    right: 0,
+    bottom: 0,
+    height: 2,
+  },
+});
+
+export default styles;
